Hoist duplicated on_error into CommandGroupExtension
The custom, rules, and sounds extensions each carried a byte-identical
on_error. Move it to the base class so every extension gets consistent
check-failure handling, resolving the logger from the subclass's module so
log attribution is unchanged.
