+++ threepseat/ext/custom/commands.py
@@ -226,15 +226,3 @@ class CustomCommands(CommandGroupExtension):
                 f'The command /{name} does not exist.',
                 ephemeral=True,
             )
-
-    async def on_error(
-        self,
-        interaction: discord.Interaction[discord.Client],
-        error: app_commands.AppCommandError,
-    ) -> None:
-        """Callback for errors in child functions."""
-        if isinstance(error, app_commands.CheckFailure):
-            await interaction.response.send_message(str(error), ephemeral=True)
-            logger.info('app command check failed: %s', error)
-        else:
-            logger.exception(error)
